Refactoring onCreate, onActivityResult, updateUI
onCreate: the signInRequest and oneTapClient implementation have been moved in there own fucntion.

onActivityResult: the failure on signing is now handle by the oneTapException function.

updateUI: when the user is checked to be non null, addNewUser is called to check in the database if the user is already stored, and if not it adds it.